+++ modules/winwave/play.c
@@ -164,7 +164,7 @@ static int write_stream_open(struct auplay_st *st,
 	wfmt.wFormatTag      = format;
 	wfmt.nChannels       = prm->ch;
 	wfmt.nSamplesPerSec  = prm->srate;
-	wfmt.wBitsPerSample  = st->sampsz * 8;
+	wfmt.wBitsPerSample  = (WORD)(st->sampsz * 8);
 	wfmt.nBlockAlign     = (prm->ch * wfmt.wBitsPerSample) / 8;
 	wfmt.nAvgBytesPerSec = wfmt.nSamplesPerSec * wfmt.nBlockAlign;
 	wfmt.cbSize          = 0;
